Bug Fix: PTAD != PTBD
A simple test code:
     while(1) PTAD = PTBD
did not work. In simulation, however, the test worked ok.
A simple loopback test (UART) did not work too.
The only workaround is add an one clock delay somewhere
(but the real reason is still unknown, probably some timing
error inside the FPGA? Maybe the arbiter?), so I removed the
conditional stall for branch in ID, for data in EX. Now, ID
will stall if a data dependency for load instruction is
detected in EX stage.

TODO: make a post-route simulation (try the memory synthesis
again).